+++ frame/lending/src/tests/vault.rs
@@ -1,6 +1,6 @@
 use crate::{
 	self as pallet_lending,
-	mocks::*,
+	mocks::general::*,
 	tests::{assert_extrinsic_event, assert_no_event, create_simple_market},
 	Error,
 };
